Make-based generation of Dockerfiles for:
1. Base OS images (32- and 64-bit) for Debian 6-9, Ubuntu 12.04 and
   14.04, and CentOS 6 and 7 (RHEL 7 is 64-bit only).
2. Wheel build images (Python 2.6 and 2.7, UCS2 and UCS4) on all those
   base OS images. Wheel build images include the evolving modification
   to the Python wheel package that I'm working on.

What this will now do:

- Build "generic" wheels that'll run on platforms at least as old as
  Debian Squeeze (so everything listed above).
- Build "targeted" wheels of things with complex/important system
  dependencies. The case for this is psycopg2, which will use the system
  libpq and have SSL support, neither of which we would then be
  responsible for keeping up to date.

+++ wheel.py
@@ -17,6 +17,7 @@ WHEELS_YML = join(WHEELS_BUILD_DIR, 'wheels.yml')
 
 def main():
     parser = argparse.ArgumentParser(description='Build wheels in Docker')
+    parser.add_argument('--image', '-i', help='Build only on this image')
     parser.add_argument('package', help='Package name (in wheels.yml)')
     args =  parser.parse_args()
 
@@ -25,10 +26,14 @@ def main():
 
     assert args.package in wheels['packages'], 'Not in %s: %s' % (WHEELS_YML, args.package)
 
-    try:
-        imageset = wheels['packages'][args.package]['imageset']
-    except:
-        imageset = 'default'
+    if args.image is not None:
+        images = [args.image]
+    else:
+        try:
+            imageset = wheels['packages'][args.package]['imageset']
+            images = wheels['imagesets'][imageset]
+        except:
+            images = wheels['imagesets']['default']
 
     src_cache = join(WHEELS_BUILD_DIR, 'cache')
     if not os.path.exists(src_cache):
@@ -37,11 +42,12 @@ def main():
     src_url = wheels['packages'][args.package]['src']
     tgz = join(src_cache, basename(src_url))
 
-    with open(tgz, 'w') as handle:
-        r = urllib2.urlopen(src_url, None, 15)
-        handle.write(r.read())
+    if not os.path.exists(tgz):
+        with open(tgz, 'w') as handle:
+            r = urllib2.urlopen(src_url, None, 15)
+            handle.write(r.read())
 
-    for image in wheels['imagesets'][imageset]:
+    for image in images:
         try:
             buildpy = wheels['images'][image]['buildpy']
         except:
@@ -49,7 +55,7 @@ def main():
         cmd = [ 'docker', 'run',
                 '--volume=%s/:/host/dist/' % WHEELS_DIST_DIR,
                 '--volume=%s/:/host/build/:ro' % WHEELS_BUILD_DIR,
-                image, buildpy, '-u', '/host/build/build.py', args.package ]
+                image, buildpy, '-u', '/host/build/build.py', args.package, image ]
         print 'Running docker:', ' '.join(cmd)
         subprocess.check_call(cmd)
